Refactor: extract shared formatters + add no-duplicate-code rule
- Created app/ui/src/utils/formatters.js with formatDate, formatValue,
  computeHistoryDiffs, and friendlyLabel, previously duplicated across
  5 components (UserDetailPage, ResourceDetailPage, GroupDetailPage,
  AccessPackageDetailPage, GovernancePage)
- Added section 4 "No Duplicate Code" to CLAUDE.md Key Conventions,
  listing known shared utilities and requiring extraction at 3+ copies

No behavior change.
